requirements.txt
PyExifTool
opencv-python
rawpy